+++ include/GafferUSD/USDShader.h
@@ -1,6 +1,6 @@
 //////////////////////////////////////////////////////////////////////////
 //
-//  Copyright (c) 2022, Cinesite VFX Ltd. All rights reserved.
+//  Copyright (c) 2023, Cinesite VFX Ltd. All rights reserved.
 //
 //  Redistribution and use in source and binary forms, with or without
 //  modification, are permitted provided that the following conditions are
@@ -36,18 +36,32 @@
 
 #pragma once
 
+#include "GafferUSD/Export.h"
+#include "GafferUSD/TypeIds.h"
+
+#include "GafferScene/Shader.h"
+
 namespace GafferUSD
 {
 
-enum TypeId
+class GAFFERUSD_API USDShader : public GafferScene::Shader
 {
 
-	USDLayerWriterTypeId = 110225,
-	USDAttributesTypeId = 110226,
-	USDShaderTypeId = 110227,
+	public :
 
-	LastTypeId = 110250,
+		explicit USDShader( const std::string &name=defaultName<USDShader>() );
+		~USDShader() override;
+
+		GAFFER_NODE_DECLARE_TYPE( GafferUSD::USDShader, USDShaderTypeId, GafferScene::Shader );
+
+		void loadShader( const std::string &shaderName, bool keepExistingValues = false ) override;
+
+	protected :
+
+		IECore::ConstCompoundObjectPtr attributes( const Gaffer::Plug *output ) const override;
 
 };
+
+IE_CORE_DECLAREPTR( USDShader )
 
 } // namespace GafferUSD