+++ python/GafferUSDTest/__init__.py
@@ -37,6 +37,7 @@
 from .ModuleTest import ModuleTest
 from .USDAttributesTest import USDAttributesTest
 from .USDLayerWriterTest import USDLayerWriterTest
+from .USDShaderTest import USDShaderTest
 
 if __name__ == "__main__":
 	import unittest